+++ src/components/MyOrders/MyOrders.js
@@ -7,6 +7,7 @@ const MyOrders = () => {
     const [myOrders, setMyOrders] = useState([]);
     const [rooms, setRooms] = useState([]);
     const [found, setFound] = useState([]);
+    const [loading, setLoading] = useState(true);
     const { user } = useAuth();
 
     useEffect(() => {
@@ -28,36 +29,43 @@ const MyOrders = () => {
     }, [])
 
     useEffect(() => {
+        setLoading(true);
         const found = [];
         rooms.forEach(room => {
             myOrders.forEach(order => {
                 if (order.order === room._id) {
                     const newObj = room;
                     newObj.orderId = order._id;
+                    newObj.status = order.status;
+                    newObj.date = order.date;
                     // console.log(newObj);
                     found.push(newObj)
                 }
             })
         })
         setFound(found);
+        setLoading(false);
     }, [rooms, myOrders])
 
     const handleCancelOrder = (id) => {
-        console.log(id);
-        // const url = `http://localhost:5000/deleteOrder/${id}`;
-        const url = `https://cryptic-temple-38934.herokuapp.com/deleteOrder/${id}`;
-        fetch(url, {
-            method: 'DELETE'
-        })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data);
-                if (data.deletedCount) {
-                    alert('successfullly canceled!')
-                    const reamainingServices = found.filter(service => service.orderId !== id)
-                    setFound(reamainingServices);
-                }
+        // console.log(id);
+        const proceed = window.confirm('Are you sure, you want to cancel the order?')
+        if (proceed) {
+            // const url = `http://localhost:5000/deleteOrder/${id}`;
+            const url = `https://cryptic-temple-38934.herokuapp.com/deleteOrder/${id}`;
+            fetch(url, {
+                method: 'DELETE'
             })
+                .then(res => res.json())
+                .then(data => {
+                    console.log(data);
+                    if (data.deletedCount) {
+                        alert('successfullly canceled!')
+                        const reamainingServices = found.filter(service => service.orderId !== id)
+                        setFound(reamainingServices);
+                    }
+                })
+        }
 
     }
 
@@ -65,7 +73,7 @@ const MyOrders = () => {
         <Container className='mt-5'>
             <Row lg={3} className="g-4">
                 {
-                    found.length > 0 ?
+                    !loading ?
                         found.map(room => <Order
                             key={room._id}
                             room={room}
